Move MemoEntry to its own header
This type will be gaining more responsibilities, should be moved to a location by itself to give it room to grow.
