+++ autowiring/CoreContext.h
@@ -17,6 +17,7 @@
 #include "InvokeRelay.h"
 #include "JunctionBoxManager.h"
 #include "member_new_type.h"
+#include "MemoEntry.h"
 #include "CoreObjectDescriptor.h"
 #include "result_or_default.h"
 #include "TeardownNotifier.h"
@@ -154,26 +155,6 @@ public:
   /// </summary>
   /// \sa AutoGlobalContext, GlobalCoreContext
   static std::shared_ptr<CoreContext> GetGlobal(void);
-
-  /// \internal
-  /// <summary>
-  /// Represents a single entry, together with any deferred elements waiting on the satisfaction of this entry
-  /// </summary>
-  struct MemoEntry {
-    // The first deferrable autowiring which requires this type, if one exists:
-    DeferrableAutowiring* pFirst = nullptr;
-
-    // A back reference to the concrete type from which this memo was generated.  This field may be null
-    // if there is no corresponding concrete type.
-    const CoreObjectDescriptor* pObjTraits;
-
-    // Once this memo entry is satisfied, this will contain the AnySharedPointer instance that performs
-    // the satisfaction
-    AnySharedPointer m_value;
-
-    // A flag to determine if this memo entry was set from the current context.
-    bool m_local = true;
-  };
 
 protected:
   // A pointer to the parent context
