+++ src/autowiring/MemoEntry.cpp
@@ -1,0 +1,5 @@
+// Copyright (C) 2012-2015 Leap Motion, Inc. All rights reserved.
+#include "stdafx.h"
+#include "MemoEntry.h"
+
+MemoEntry::MemoEntry(void) {}
